+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ### Added
 - Managed Database OpenSearch support
+- Support for defining NIC model upon creating or modifying a server. Also exported constants for each support NIC model.
 
 ## [6.1.1]