+++ CHANGELOG.md
@@ -8,6 +8,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Managed Database OpenSearch support
 - Support for defining NIC model upon creating or modifying a server. Also exported constants for each support NIC model.
+- Support for `PrivateNodeGroups` property of Kubernetes clusters.
 
 ## [6.1.1]