+++ CHANGELOG.md
@@ -8,6 +8,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Managed Database OpenSearch support
 - Support for defining NIC model upon creating or modifying a server. Also exported constants for each support NIC model.
+- Support for `PrivateNodeGroups` property of Kubernetes clusters.
 
 ### Changed
 - client: overwrite the HTTP Client Transport accordingly when `UPCLOUD_DEBUG_SKIP_CERTIFICATE_VERIFY` is set to `1`